move satf converts to the converter folder
